# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RFID API Server",
    "version": "1.0.0",
    "description": "API for RFID Reader integration with JWT Auth and User Management"
  },
  "paths": {
    "/login": {
      "post": {
        "summary": "User Login",
        "description": "Authenticate user and get JWT tokens.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": { "type": "string" },
                    "refresh_token": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },

    "/refresh": {
      "post": {
        "summary": "Refresh Access Token",
        "description": "Refresh access token using refresh token.",
        "security": [
          { "bearerAuth": [] }
        ],
        "responses": {
          "200": {
            "description": "New access token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },

    "/api/rfid": {
      "post": {
        "summary": "Send RFID Data",
        "description": "Submit RFID EPC and RSSI data from reader.",
        "security": [
          { "bearerAuth": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "epc": { "type": "string" },
                  "rssi": { "type": "integer" },
                  "ipaddress": { "type": "string" },
                  "client": { "type": "string" }
                },
                "required": ["epc", "rssi", "ipaddress", "client"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Data received",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },

    "/api/rfid-logs": {
      "get": {
        "summary": "Get RFID Logs",
        "description": "Retrieve list of RFID scan logs.",
        "security": [
          { "bearerAuth": [] }
        ],
        "responses": {
          "200": {
            "description": "List of logs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "integer" },
                      "epc": { "type": "string" },
                      "rssi": { "type": "integer" },
                      "ipaddress": { "type": "string" },
                      "client": { "type": "string" },
                      "timestamp": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },

    "/api/users": {
      "get": {
        "summary": "List Users",
        "description": "Get list of all users.",
        "security": [
          { "bearerAuth": [] }
        ],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "integer" },
                      "username": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        }
      },

      "post": {
        "summary": "Add New User",
        "description": "Create a new user.",
        "security": [
          { "bearerAuth": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },

    "/api/users/{username}": {
      "put": {
        "summary": "Update User Password",
        "description": "Change password for a user.",
        "security": [
          { "bearerAuth": [] }
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": { "type": "string" }
                },
                "required": ["password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },

      "delete": {
        "summary": "Delete User",
        "description": "Delete a user by username.",
        "security": [
          { "bearerAuth": [] }
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    }
  },

  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}